Move relevant kernel interop tests into kernel_tests.
